+++ src/geom/shape.rs
@@ -1,4 +1,4 @@
-use crate::geom::{Circle, Line, Rectangle, Triangle, Vector, about_equal};
+use crate::geom::{Circle, Line, Rectangle, Triangle, Vector, Transform, about_equal};
 
 /// The collision and positional attributes of shapes
 pub trait Shape {
@@ -24,6 +24,25 @@ pub trait Shape {
     /// A Rectangle that contains the entire shape
     #[must_use]
     fn bounding_box(&self) -> Rectangle;
+    /// Apply a transform to a shape then get the bounding box for the transformed shape
+    #[must_use]
+    fn transform_bounding_box(&self, transform: Transform) -> Rectangle {
+        let bb = self.bounding_box();
+        // Build the transform to rotate the shape
+        let transform = Transform::translate(bb.center() + bb.pos)
+            * transform
+            * Transform::translate(-bb.pos - bb.center());
+        // Get new corner position
+        let tl = transform * bb.pos;
+        let tr = transform * Vector::new(bb.pos.x + bb.size.x, bb.pos.y);
+        let bl = transform * Vector::new(bb.pos.x, bb.pos.y + bb.size.y);
+        let br = transform * Vector::new(bb.pos.x + bb.size.x, bb.pos.y + bb.size.y);
+        // Get min and max points
+        let min = tr.min(tl).min(bl).min(br);
+        let max = tr.max(tl).max(bl).max(br);
+        // Make new bounding box
+        Rectangle::new(min, max - min)
+    }
     /// Create a copy of the shape with an offset center
     #[must_use]
     fn translate(&self, amount: impl Into<Vector>) -> Self where Self: Sized;